Add file metadata

I unfortunately made this PR far too big. Started refactoring a fair bit so it's a mess.

Main changes:

- Added a "metadata" for each table/file with snippets of relevant article text
- Added the beginning of an impl to extract extra metadata from supp files
- Added type hinting throughout
- Updated abstract base classes for article managers, lists, and threads

+++ app/model/file_io.py
@@ -2,27 +2,50 @@ import pandas as pd
 import requests
 import os
 
-def extract_dfs(fname): # returns a dictionary of dataframes
+
+def extract_dfs(fname: 'str', should_stop: 'bool') -> 'dict[str, pd.DataFrame]':
+    if should_stop:
+        return
+
     if fname.endswith(('.xlsx')):
-        data = pd.read_excel(fname, sheet_name=None, header=None, index_col=None)
+        if should_stop:
+            return
+        data = pd.read_excel(
+            fname, sheet_name=None, header=None, index_col=None
+        )
     elif fname.endswith('.csv'):
+        if should_stop:
+            return
         data = {"sheet": pd.read_csv(fname, header=None, index_col=None)}
     else:
-        data = {"sheet": pd.read_csv(fname, delimiter='\t', header=None, index_col=None)}
-    
+        if should_stop:
+            return
+        data = {"sheet": pd.read_csv(
+            fname, delimiter='\t', header=None, index_col=None)}
+
+    if should_stop:
+        return
+
     os.remove(fname)
 
     return data
 
 
-def download_supp(url):
+def download_supp(url: 'str', should_stop: 'str') -> 'str':
+    if should_stop:
+        return
+
     local_fname = url.split('/')[-1]
     if not local_fname.endswith(('.txt', '.csv', '.xlsx')):
         return
+
     with requests.get(url, stream=True) as r:
         r.raise_for_status()
         with open(local_fname, 'wb') as f:
             for chunk in r.iter_content(chunk_size=8192):
+                if should_stop:
+                    os.remove(local_fname)
+                    return
                 f.write(chunk)
 
     return local_fname
